Add 'type' argument to argument parser
This commit adds a new argument to the argument parser. The argument
is called 'type' and is used to specify the type of log file that is
being parsed.
If 'type' is not specified, the argument parser will try to guess the
type of the log file based on the file extension.

Signed-off-by: GitHub <noreply@github.com>